+++ packages/core/modules/component/component-class.js
@@ -87,7 +87,7 @@ class Framework7Component {
     if (html && typeof html === 'string') {
       html = html.trim();
       self.$vnode = vdom(html, self, app, true);
-      self.el = document.createElement('div');
+      self.el = document.createElement(self.$vnode.sel || 'div');
       patch(self.el, self.$vnode);
     } else if (html) {
       self.el = html;