+++ packages/react/components/fab-button.js
@@ -100,6 +100,22 @@ class F7FabButton extends React.Component {
   componentDidUpdate(prevProps, prevState) {
     __reactComponentWatch(this, 'props.tooltip', prevProps, prevState, newText => {
       const self = this;
+
+      if (!newText && self.f7Tooltip) {
+        self.f7Tooltip.destroy();
+        self.f7Tooltip = null;
+        delete self.f7Tooltip;
+        return;
+      }
+
+      if (newText && !self.f7Tooltip && self.$f7) {
+        self.f7Tooltip = self.$f7.tooltip.create({
+          targetEl: self.refs.el,
+          text: newText
+        });
+        return;
+      }
+
       if (!newText || !self.f7Tooltip) return;
       self.f7Tooltip.setText(newText);
     });