+++ packages/vue/components/list-group.js
@@ -6,7 +6,11 @@ export default {
   props: Object.assign({
     id: [String, Number],
     mediaList: Boolean,
-    sortable: Boolean
+    sortable: Boolean,
+    sortableMoveElements: {
+      type: Boolean,
+      default: undefined
+    }
   }, Mixins.colorProps),
 
   render() {
@@ -18,7 +22,8 @@ export default {
       id,
       style,
       mediaList,
-      sortable
+      sortable,
+      sortableMoveElements
     } = props;
     const classes = Utils.classNames(className, 'list-group', {
       'media-list': mediaList,
@@ -28,7 +33,8 @@ export default {
       style: style,
       class: classes,
       attrs: {
-        id: id
+        id: id,
+        'data-sortable-move-elements': typeof sortableMoveElements !== 'undefined' ? sortableMoveElements.toString() : undefined
       }
     }, [_h('ul', [this.$slots['default']])]);
   },