+++ packages/vue/framework7-vue.esm.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 Vue 4.4.10
+ * Framework7 Vue 4.5.0
  * Build full featured iOS & Android apps using Framework7 & Vue
  * http://framework7.io/vue/
  *
@@ -7,7 +7,7 @@
  *
  * Released under the MIT License
  *
- * Released on: July 29, 2019
+ * Released on: August 21, 2019
  */
 
 import f7AccordionContent from './components/accordion-content';